% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_packages.R
\name{download_packages}
\alias{download_packages}
\title{Download packages locally}
\usage{
download_packages(pkgs_to_download, typefiles)
}
\arguments{
\item{pkgs_to_download}{character list with packages to download}

\item{typefiles}{string indicating what type to download (source or binary)}
}
\value{
the directory ./pgk_source_files is created if it does not exists and
  all the files are downloaded to this directory.
}
\description{
Download packages locally
}
\examples{
download_packages("ggplot2","source")
}
