% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Get dependencies for package list to be installed}
\usage{
get_dependencies(pkgs_to_download)
}
\arguments{
\item{pkgs_to_download}{list with packages}
}
\description{
Function necessary for the \code{install_updates} and
  \code{install_newpackages} function. It finds the dependencies for those
  packages that are to be updated or installed.
}
\examples{
get_dependencies(c("ggplot2", "dplyr"))
}
