% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_updates.R
\name{install_updates}
\alias{install_updates}
\title{Install updates from local repositories}
\usage{
install_updates(packlist, typefiles = "source")
}
\arguments{
\item{packlist}{list with packages to be updated}

\item{typefiles}{string either "source" (tar.gz) or "binary" (zip)}
}
\description{
Install updates from local repositories
}
\examples{
# install_updates("ggplot2", "source")
}
