% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_repo_dir.R
\name{set_repo_dir}
\alias{set_repo_dir}
\title{Set repositories to look for packages}
\usage{
set_repo_dir()
}
\value{
character vector of repositories
}
\description{
Set repositories to look for packages
}
\examples{
repositories <- set_repo_dir()
}
