% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_newpackages.R
\name{install_newpackages}
\alias{install_newpackages}
\title{Install list of packages from local directory}
\usage{
install_newpackages(listtoinstall, typefiles = "source")
}
\arguments{
\item{listtoinstall}{List of packages to install}

\item{typefiles}{string either source (tar.gz) or binary (zip)}
}
\description{
Install list of packages from local directory
}
