% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_updates.R
\name{install_updates}
\alias{install_updates}
\title{Install updates from local repositories}
\usage{
install_updates(typefiles)
}
\arguments{
\item{typefiles}{string either "source" (zip) or "binary" (tar.gz)}
}
\description{
Install updates from local repositories
}
\examples{
install_updates("source")
}
